% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cicero_data}
\alias{cicero_data}
\title{Example single-cell chromatin accessibility data}
\format{
A data frame with 35137 rows and 3 variables:
\describe{
  \item{Peak}{Peak information}
  \item{Cell}{Cell ID}
  \item{Count}{Reads per cell per peak}
}
}
\usage{
cicero_data
}
\description{
A dataset containing a subset of a single-cell ATAC-seq
dataset collected on Human Skeletal Muscle Myoblasts.
Only includes data from chromosome 18.
}
\keyword{datasets}
