% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df_for_coords}
\alias{df_for_coords}
\title{Construct a data frame of coordinate info from coordinate strings}
\usage{
df_for_coords(coord_strings)
}
\arguments{
\item{coord_strings}{A list of coordinate strings (each like
"chr1:500000-1000000")}
}
\description{
Construct a data frame of coordinate info from coordinate strings
}
\details{
Coordinate strings consist of three pieces of information:
  chromosome, start, and stop. These pieces of information can be separated
  by the characters ":", "_", or "-". Commas will be removed, not used as
  separators (ex: "chr18:8,575,097-8,839,855" is ok).
}
