% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_annotation_sample}
\alias{gene_annotation_sample}
\title{Example gene annotation information}
\format{A data frame with 15129 rows and 8 variables:
\describe{
  \item{chromosome}{Chromosome}
  \item{start}{exon starting base}
  \item{end}{exon ending base}
  \item{strand}{exon mapping direction}
  \item{feature}{feature type}
  \item{gene}{gene ID}
  \item{transcript}{transcript ID}
  \item{symbol}{gene symbol}
}}
\usage{
gene_annotation_sample
}
\description{
Gencode gene annotation data from chromosome 18 of the
human genome (hg19).
}
\keyword{datasets}
