% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human.hg19.genome}
\alias{human.hg19.genome}
\title{Chromosome lengths from human genome hg19}
\format{
A data frame with 93 rows and 2 variables:
\describe{
  \item{V1}{Chromosome}
  \item{V2}{Chromosome length, base pairs}
}
}
\usage{
human.hg19.genome
}
\description{
A list of the chromosomes in hg19 and their lengths
in base pairs.
}
\keyword{datasets}
