% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_sparse_matrix}
\alias{make_sparse_matrix}
\title{Make a symmetric square sparse matrix from data frame}
\usage{
make_sparse_matrix(data, i.name = "Peak1", j.name = "Peak2", x.name = "value")
}
\arguments{
\item{data}{data frame}

\item{i.name}{name of i column}

\item{j.name}{name of j column}

\item{x.name}{name of value column}
}
\value{
sparse matrix
}
\description{
Convert a data frame into a square sparse matrix (all versus all)
}
