% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse.mm9.genome}
\alias{mouse.mm9.genome}
\title{Chromosome lengths from mouse genome mm9}
\format{
A data frame with 93 rows and 2 variables:
\describe{
  \item{V1}{Chromosome}
  \item{V2}{Chromosome length, base pairs}
}
}
\usage{
mouse.mm9.genome
}
\description{
A list of the chromosomes in mm9 and their lengths
in base pairs.
}
\keyword{datasets}
