% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activityScores.R
\name{normalize_gene_activities}
\alias{normalize_gene_activities}
\title{Normalize gene activities}
\usage{
normalize_gene_activities(activity_matrices, cell_num_genes)
}
\arguments{
\item{activity_matrices}{A gene activity matix, output from
\code{\link{build_gene_activity_matrix}}, or a list of gene activity
matrices to be normalized together.}

\item{cell_num_genes}{A named vector of the total number of accessible sites
per cell. Names should correspond to the cell names in the activity
matrices. These values can be found in the "num_genes_expressed" column of
the pData table of the CDS used to calculate the gene activity matrix.}
}
\value{
Normalized activity matrix or matrices.
}
\description{
Normalize the output of \code{\link{build_gene_activity_matrix}}. Input is
either one or multiple gene activity matrices. Any gene activties to be
compared amongst each other should be normalized together.
}
\examples{
\dontrun{
num_genes <- pData(input_cds)$num_genes_expressed
names(num_genes) <- row.names(pData(input_cds))
cicero_gene_activities <- normalize_gene_activities(unnorm_ga, num_genes)
}


}
