% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cicero.R
\docType{package}
\name{cicero-package}
\alias{cicero}
\alias{cicero-package}
\title{cicero}
\description{
Cicero computes putative cis-regulatory maps from single-cell chromatin accessibility data. It also extends monocle 2 for use in chromatin accessibility data.
}
\author{
\strong{Maintainer}: Hannah Pliner \email{hpliner@uw.edu}

Authors:
\itemize{
  \item Cole Trapnell \email{coletrap@uw.edu}
}

}
