% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_annotation_sample}
\alias{gene_annotation_sample}
\title{Example gene annotation information}
\format{
A data frame with 15129 rows and 8 variables:
\describe{
  \item{chromosome}{Chromosome}
  \item{start}{Exon starting base}
  \item{end}{Exon ending base}
  \item{strand}{Exon mapping direction}
  \item{feature}{Feature type}
  \item{gene}{Gene ID}
  \item{transcript}{Transcript ID}
  \item{symbol}{Gene symbol}
}
}
\usage{
gene_annotation_sample
}
\description{
Gencode gene annotation data from chromosome 18 of the
human genome (hg19).
}
\keyword{datasets}
